+++ src/fractal_client/cmd/__init__.py
@@ -2,7 +2,7 @@ from httpx2 import Client
 from httpx2 import ConnectError
 
 from fractal_client import __VERSION__
-from fractal_client.auth import AuthClient
+from fractal_client.auth._client import AuthClient
 from fractal_client.interface import Interface
 
 from ._dataset import delete_dataset
